test: 코인 출력 테스트 작성 및 기존 파일 삭제
Test코드 작성법에 대한 미숙으로 기존 파일을 삭제하고 Application.kt로 내용 옮김
